:adhesive_bandage: changement du type de l'attribut EXPLOSION_DURATION

+++ src/main/java/fr/univartois/butinfo/r304/bomberman/model/movables/bomb/Explosion.java
@@ -10,7 +10,7 @@ import java.util.Objects;
 
 public class Explosion extends AbstractMovable {
 
-    public static final double EXPLOSION_DURATION = 500; // 1/2 seconde
+    public static final long EXPLOSION_DURATION = 500; // 1/2 seconde
     private final long explosionBegin;
     public static SpriteStore spriteStore = new SpriteStore();
